requirements.txt
numpy==1.14.3
tensorflow-gpu==1.8.0
keras==2.2.0
scikit-learn==0.19.1
matplotlib==2.2.2